+++ src/connectField.js
@@ -1,11 +1,8 @@
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
-
+import get from 'lodash/get'
 import { getActions, getState } from './index'
 
-function _selectForm(state) {
-  return state.form
-}
 // This gets state and actions for a specific field. That is all.
 // The formId and fieldId can be sent via init as an object or with props on each instance.
 export default function connectField(options = {}) {
@@ -13,8 +10,8 @@ export default function connectField(options = {}) {
   function getInfo(ownProps) {
     return {
       formId: ownProps.formId || options.formId || 'default',
-      fieldId: ownProps.field.id || ownProps.fieldId || options.fieldId || 'NO_FIELD_ID',
-      validate: ownProps.field.validate,
+      fieldId: ownProps.fieldId || get(ownProps, 'field.id', options.fieldId) || 'NO_FIELD_ID',
+      validate: get(ownProps, [ 'field', 'validate' ]),
     }
   }
   // Pass in a component and it will get connected for you.