+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [2.0.0]
+> 2016-02-28
+
+* Actions are now split into groups. `fieldEvent` contains all actions related to field state. `formEvent` are form and focus events like `onChange`. `formHandler` a clone of `formEvent` where the keys replace `on` with `handle` like `handleBlur` instead of `onBlur`.
+* 
+
+## [1.4.0]
+> 2016-02-23
+
+* reducer on `SAVED` will hoist a payload `id` prop to the field state. It will be left as-is to be saved to `savedValue`. The purpose is to allow the server to set a unique field id.
+* `saved()` action now accepts payload and value.
+* Update `value` on `BLUR` if one is given.
+
 ## [1.3.0]
 > 2016-02-19